Image testing
I was able to test out if the user enters a valid SKU then the image will pop up in the picture box (named pbProduct). It is a success however when I went to test a second time Visual Studio Locked me out. I am currently trying to resolve the problem