Reset Password Code
I am starting to code for the Reset/ update password screen. So far I have only added a few controls for verifying the random code that will be sent via email. The controls are hidden when the user first enters the form and will be shown once the update password button is clicked